requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
waitress==1.4.4
Werkzeug==1.0.1
pyOpenSSL==19.0.0
